inference/docker/requirements.txt
pandas
transformers
tensorflow
psycopg2-binary